September 30, 2025
First back up save file to Github. Placeholder sprites for objects and tileset, along with basic movement and a attack button. I'm having troubles on getting the player to stop during the "attack state" and fully play the correct animation along with the enemy checking for the players distance to start chasing.