October 3, 2025
Fixed enemy tracking and added a new room to figure out how to change rooms with doors and saving the ghosts you killed by changing the rooms to be 'persistent'. I'll need to work on two global variables; one being the point system for every time you kill a ghost and the other is the players HP bar since it won't work with being built into the player object.