October 7, 2025
I fixed the text on the start screen as well as attempting to make a global score variable. It shows the score at the top left and does add points to the score when you kill the enemy but just like the health bar, it resets when entering a new room.