October 8, 2025
Fixed the health bar and score system so they don't reset when entering a new room. Also put the health and score system into one "GameManager" object for now. I want to organize it better later so it's easier to look at (The health system is fine to look at but the score isn't).